More PR 41 suggestions.
